// src/app/models/doi/doiModel.ts
export class MasterEntry {
    schemeName: string;
    schemeId: number;
    schemeNameGuj: string;
    schemeShortname: string;
    nodalOffice: string;
    nodelOfficeId: number;
    schemeStatusId: number;
    claimAmount: number;
    createdDate: string;
    updatedDate: string;
    status: number;
    maximumAgeRange: number;
    minimumAgeRange: number;
    remarks: string;
    referenceNo: string;
    referenceDate: string;
    activeStatus: number;
}